+++ src/app/locales/en/translations.js
@@ -506,6 +506,7 @@ export default {
       providersMapOfDist: 'Active transfers map',
       activeTransfers: 'Active transfers',
       completedTransfers: 'History of transfers',
+      onTheFlyTransfers: 'On-the-fly transfers',
       noActiveTransfers: 'There are no active transfers',
       noCompletedTransfers: 'There are no past transfers',
       initializingTransfers: 'Initializing transfers...',
@@ -551,6 +552,7 @@ export default {
         throughput: 'Throughput',
         output: 'Output',
         waitingForTransferStart: 'Waiting for the transfer to start...',
+        noStatsForUnit: 'There are no transfer statistics for the last {{timeUnit}}.',
         waitingForStats: 'Gathering transfer statistics...',
         waitingForStatsTip: 'Statistics are delayed due to synchronization ' +
           'latency caused by data distribution.',
@@ -561,6 +563,9 @@ export default {
         output: 'Output',
         timeLastUpdate: 'Time (last update: {{lastUpdate}})',
         throughput: 'Throughput',
+        onTheFly: 'On-the-fly',
+        all: 'All',
+        jobs: 'Jobs',
       },
     },
     rootDirSettingsDrop: {